Combined default and non-default share creation into one template function (#590)
* Improved replication documentation

* Replication: Added active/passive documentation

* Fixed typo

* Added see also section at the end

* Added Peer Persistence based replication documentation

* Missed out Peer Persistence based replication documentation in last commit

* Increased title font for PP based replication documentation

* Added a note

* Introductory content updated for PP documentation

* Added content related to few more restrictions

* Updated a restriction with more details

* Fix for #428

* Revert "Fix for #428"

This reverts commit f074ae3df7e0459214c2652379ead5ce3e440abd.

* Fix for issue #428

Covered following TCs:
1. With only QOS
2. With only flash-cache
3. With both
4. Without both i.e. just a VVSet with the source volume member of it

* File Persona Support

This is work in progress
So far implemented:
* CRD operations
* Share state management

TODO:
* Rollback requires some work
* Testing of some scenarios

* File Persona: using single configuration file

Implemented the following:
==================
1. Dependency on common configuration file between block and file protocols
2. Adding of client IP access via WSAPI call

TODOs:
=====
1. Unit test implementation to adapt to share creation on child thread. Presently it fails.
2. Rollback
3. Quota size
4. Testing of some scenarios

* Fixed typo in function name

* Fixed PEP8 issues

* Commented out fix for issue #428 for now

* Fixed UT failures

Due to changes to the design, block UTs were failing. Fixed those.

* Fixed couple of more PEP8 issues

* Added code for multiple default backends

* Expect cpg to be list type in hpe.conf

In block, cpg is a list type in hpe.conf. File earlier used expect cpg to be string type. After common configuration file, File needed this change

* Fixed broken Travis CI

* Fixed unit test related to listing of volumes

*Cannot rely on first manager anymore as user may or may not configure both the managers.

* Fixed multiple issues

Implemented following:
1. IP range
2. Delete FPG with last share delete
3. Renamed "persona" flag to "filePersona"
4. Fixed mount/unmount
5. Fixed default share size
6. Lock by share name
7. In share meta-data, IP/Subnet were not getting updated for second share onwards

* Update file_backend_orchestrator.py

Added one missing paramter

* Fixed mount/unmount + Addressed review comment

* Mount infomration needed to be stored as a dictionary with mount_id as key and mount_dir as value
* If default FPG dict is empty, needed to throw exception EtcdDefaultFpgNotPresent
* Removed replication related code

* Update file_manager.py

Fixed couple of PEP8 issues

* Update hpe_3par_mediator.py

Fixed the configuration parameter names

* Review Comments addressed

* Unit test framework fixed for decrypt_password

* Rollback for default share creation

TODO:
* Rollback for non-default share creation

* Resolved PEP8 errors

* Fixed async initialization failure in UTs

* Update cmd_deleteshare.py

Fixed typo

* Update cmd_deleteshare.py

Fixed typo

* Added logging

* Backend metadata initialization done for a use case

* PEP8 fixed + Quota set in a use case

* Combined default and non-default share creation into one template function

create_share_on_fpg() can now create both default and non-default shares.
It achieves this by using specific implementations of two functions the references of which are passed to it as argument. Rest of the function steps are common to both default and non-default share creation process.

* Removed unused import references

* Async initialization fix

* Added sleep to main UT thread

* Fixed PEP8

* Default size modified + share create on legacy FPG issue fixed

*Changed default FPG to 16TiB and default share size to 1TiB
*Fixed share creation issue on legacy FPG

* Updated error message for legacy FPG use case

+++ hpedockerplugin/cmd/cmd_deleteshare.py
@@ -25,12 +25,14 @@ class DeleteShareCmd(cmd.Cmd):
                 self._backend, self._cpg_name, self._fpg_name):
             self._remove_quota()
             self._delete_share()
-            remaining_cnt = self._decrement_share_cnt()
-            if remaining_cnt == 0:
-                self._delete_fpg()
+            # Decrement count only if it is Docker managed FPG
+            if self._share_info.get('docker_managed'):
+                remaining_cnt = self._decrement_share_cnt()
+                if remaining_cnt == 0:
+                    self._delete_fpg()
         return json.dumps({u"Err": ''})
 
-    def _unexecute(self):
+    def unexecute(self):
         pass
 
     def _remove_quota(self):
